+++ cmd/state-tray/internal/libc-compatwrap/compatwrap_linux.go
@@ -1,5 +1,0 @@
-package compatwrap
-
-import (
-	"C"
-)