+++ internal/testhelpers/tagsuite/tagsuite.go
@@ -53,6 +53,7 @@ const (
 	Shell          = "shell"
 	Exec           = "exec"
 	Show           = "show"
+	Uninstall      = "uninstall"
 	Update         = "update"
 	VSCode         = "vscode"
 )